FocalEngine GLFW branch finished

+++ FocalEngine-dev/include/shaderManager.h
@@ -14,7 +14,7 @@ private:
 	ShaderManager(Window* window);
 	static ShaderManager* s_instance;
 	Window* m_curWindow;
-	double m_deltatime;
+	float m_deltatime;
 
 	std::unordered_map<std::string, std::shared_ptr<Shader>> m_shaderMap;
 	Shader* find(std::string componentName, std::unordered_map<std::string, std::shared_ptr<Shader>> map);
@@ -32,7 +32,7 @@ public:
 	void addShader(std::string componentName, const std::string& fragmentShader, const std::string& vertexShader);
 	Shader* getShader(std::string componentName);
 
-	void update(double* deltatime);
+	void update(float* deltatime);
 
 	double getDeltatime();
